+++ heart-rate-chicken-embryos-sdk/src/main/java/com/imukstudio/heartrate/chicken/embryos/sdk/interactor/impl/MeasureInteractorImpl.kt
@@ -17,7 +17,7 @@ class MeasureInteractorImpl(
         measureHandler.startMeasure(activity = activity, surfaceView = surfaceView, textureView = textureView)
     }
 
-    override fun subscribeMeasureResult(callback: (pulse: Int, cycles: Int, time: Float) -> Unit) {
+    override fun subscribeMeasureResult(callback: (pulse: Int, cycles: Int, time: Long) -> Unit) {
         listenersSDK.subscribeMeasureResult(callback)
     }
